azure-functions/requirements.txt
azure-functions>=1.11.0
azure-cosmos
httpx
pyyaml